refactor: 베팅머니 데이터 타입 long으로 수정

+++ src/main/java/domain/BettingMoney.java
@@ -4,17 +4,17 @@ import domain.constant.Result;
 import domain.game.ProceedsCalculator;
 
 public class BettingMoney {
-    private static final int MIN_BETTING_MONEY = 1;
+    private static final long MIN_BETTING_MONEY = 1L;
     private static final String INVALID_MESSAGE = "배팅 금액은 0보다 커야 합니다.";
 
-    private final int value;
+    private final long value;
 
-    public BettingMoney(int value) {
+    public BettingMoney(long value) {
         validate(value);
         this.value = value;
     }
 
-    private void validate(int value) {
+    private void validate(long value) {
         if (value < MIN_BETTING_MONEY) {
             throw new IllegalArgumentException(INVALID_MESSAGE);
         }
